iterate through combinations via indices
when I have M points, I know I have choose(M, 3) triangles to construct.
but before this, I did not know how to iterate through these triangles
apart from a triple-for loop.

the function invert_combi gives me the ith combination in choose(n, 3),
in lexicographic order. the python version has asserts to confirm
correctness of this method, and is equivalent to the C++ version. I
wonder if there is a O(1) way to calculate it.

the benefit of iterating like this is that OpenMP likes such for loops.