search twice by using heuristic clique for pruning
use_heuristic is usually fast (polynomial time), but may not give the
largest clique. use_dfs is slow in the worst case (exponential time),
and requires a small/sparse graph to be somewhat fast. so we run the
heuristic method first, obtain a clique, use that to strip the graph
down as much as possible (k-core type calculation or similar, noting
that core number of a graph is always greater than the clique number)
and then run the dfs search on the (hopefully) smaller graph.